backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
pgvector==0.2.4
weaviate-client==4.4.4
openai==1.3.7
sentence-transformers>=2.7.0
beautifulsoup4==4.12.2
scrapy==2.11.1
selenium==4.15.2
requests==2.31.0
schedule==1.2.1
python-multipart==0.0.6
pydantic==2.4.2
